minor correction => handlers for ICommand & IQuery had been swapped
